Engineering - move runOnChange and runOnChangeWithStore utility functions
